pyproject.toml
[tool.poetry]
name = "dji_parse"
version = "0.1.0"
description = "A simple utility for parsing telemetry data from MP4 files generated by DJI drones. Outputs data as CSV, JSON, or GPX."
authors = ["Fergus In London <fergus@fergus.london>"]
readme = "README.md"

[tool.poetry.scripts]
dji_parse = "dji_parse:run"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.3"
gpxpy = "^1.5.0"
pysrt = "^1.1.2"
ffmpeg-python = "^0.2.0"
rich = "^12.6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
